remove redundant banners
