+++ src/components/Layout/Header/NavBar/NavBar.tsx
@@ -5,12 +5,12 @@ import { useTranslate } from 'react-polyglot'
 import { Link as ReactRouterLink } from 'react-router-dom'
 import { Route } from 'Routes/helpers'
 import { routes } from 'Routes/RoutesCommon'
+import { YatBanner } from 'components/Banners/YatBanner'
 import { Text } from 'components/Text'
 import { usePlugins } from 'context/PluginProvider/PluginProvider'
 import { useFeatureFlag } from 'hooks/useFeatureFlag/useFeatureFlag'
 
 import { MainNavLink } from './MainNavLink'
-import { YatBanner } from './YatBanner'
 
 type NavBarProps = {
   isCompact?: boolean
